very minor fixes
1. Added jobname
2. Turned off trimming in concat.py
3. Corrected assert in main.py

+++ scripts/tfsemb_concat.py
@@ -52,7 +52,6 @@ def parse_arguments():
     parser.add_argument("--suffix", type=str, default="")
     parser.add_argument("--verbose", action="store_true", default=False)
     parser.add_argument("--subject", type=str, default="625")
-    parser.add_argument("--history", action="store_true", default=False)
     parser.add_argument("--conversation-id", type=int, default=0)
     parser.add_argument("--pkl-identifier", type=str, default=None)
     parser.add_argument("--project-id", type=str, default=None)
@@ -116,7 +115,11 @@ def main():
         "base_df.pkl",
     )
     os.makedirs(os.path.dirname(dst), exist_ok=True)
-    shutil.move(src, dst)
+    if os.path.exists(dst):
+        print("Base DataFrame Exists")
+    else:
+        print("Moving Base DataFrame")
+        shutil.move(src, dst)
 
     layer_folders = sorted(os.listdir(args.output_dir))
     for layer_folder in layer_folders:
@@ -149,22 +152,23 @@ def main():
         os.makedirs(os.path.dirname(full_emb_out_file), exist_ok=True)
         save_pickle(all_exs, full_emb_out_file)
 
-        if "glove" in args.embedding_type or layer_folder in [
-            "layer_48",
-            "layer_16",
-            "layer_8",
-        ]:
-            trimmed_df = load_pickle(trimmed_labels, key="labels")
-            all_df.set_index(["conversation_id", "index"], inplace=True)
-            trimmed_df.set_index(["conversation_id", "index"], inplace=True)
-            all_df["adjusted_onset"] = None
-            all_df["adjusted_offset"] = None
-            all_df.update(trimmed_df)  # merge
-            all_df.dropna(subset=["adjusted_onset"], inplace=True)
-            all_df.reset_index(inplace=True)
-            all_exs = all_df.to_dict("records")
-            fn = args.emb_out_file.replace("full", "trimmed")
-            save_pickle(all_exs, os.path.join(args.emb_out_dir, fn))
+        if False:
+            if "glove" in args.embedding_type or layer_folder in [
+                "layer_48",
+                "layer_16",
+                "layer_8",
+            ]:
+                trimmed_df = load_pickle(trimmed_labels, key="labels")
+                all_df.set_index(["conversation_id", "index"], inplace=True)
+                trimmed_df.set_index(["conversation_id", "index"], inplace=True)
+                all_df["adjusted_onset"] = None
+                all_df["adjusted_offset"] = None
+                all_df.update(trimmed_df)  # merge
+                all_df.dropna(subset=["adjusted_onset"], inplace=True)
+                all_df.reset_index(inplace=True)
+                all_exs = all_df.to_dict("records")
+                fn = args.emb_out_file.replace("full", "trimmed")
+                save_pickle(all_exs, os.path.join(args.emb_out_dir, fn))
 
     # Deleting embeddings after concatenation
     shutil.rmtree(args.output_dir, ignore_errors=True)
